Create Microsoft.DebugEngineHost
This checkin contains the initial creation of Microsoft.DebugEngineHost. This will contain the code which is specific to Visual Studio vs. Visual Studio Code hopefully allowing a debug engine to be used in both places (just with different implementions of Microsoft.DebugEngineHost.

At this point I made the MIDebugEngine reference this dll, and I moved the first set of code from MIEngine->DebugEngineHost.
