+++ src/MIDebugEngine/AD7.Impl/AD7Engine.cs
@@ -4,7 +4,6 @@
 using System;
 using System.Collections.Generic;
 using System.Text;
-using System.Runtime.InteropServices;
 using System.Runtime.ExceptionServices;
 using Microsoft.VisualStudio.Debugger.Interop;
 using System.Diagnostics;
@@ -30,8 +29,8 @@ namespace Microsoft.MIDebugEngine
     //
     // IDebugEngineProgram2: This interface provides simultanious debugging of multiple threads in a debuggee.
 
-    [ComVisible(true)]
-    [Guid("0fc2f352-2fc1-4f80-8736-51cd1ab28f16")]
+    [System.Runtime.InteropServices.ComVisible(true)]
+    [System.Runtime.InteropServices.Guid("0fc2f352-2fc1-4f80-8736-51cd1ab28f16")]
     sealed public class AD7Engine : IDebugEngine2, IDebugEngineLaunch2, IDebugProgram3, IDebugEngineProgram2, IDebugMemoryBytes2, IDebugEngine110
     {
         // used to send events to the debugger. Some examples of these events are thread create, exception thrown, module load.
